backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# LLM & Embeddings
litellm==1.17.9
openai==1.10.0
tiktoken==0.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication
google-auth==2.27.0
python-jose[cryptography]==3.3.0

# CORS & WebSocket support
websockets==12.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
